Refactor useRoomConnection
To align with implementation of `useLocalMedia`, the following is done:

- Move external reducer back into useRoomConnection module
- Change return type of useRoomConnection to be an object, not an array
- Initialize RoomConnection directly in useState, to avoid null value
