--- src/lib/reducer.ts
@@ -1,142 +1,0 @@
-import { LocalParticipant, RemoteParticipant } from "./RoomParticipant";
-
-type RemoteParticipantState = Omit<RemoteParticipant, "updateStreamState">;
-
-export interface RoomState {
-    localParticipant?: LocalParticipant;
-    roomConnectionStatus?: "connecting" | "connected" | "disconnected";
-    remoteParticipants: RemoteParticipantState[];
-}
-
-type Action =
-    | {
-          type: "ROOM_JOINED";
-          payload: {
-              localParticipant: LocalParticipant;
-              remoteParticipants: RemoteParticipantState[];
-          };
-      }
-    | {
-          type: "PARTICIPANT_AUDIO_ENABLED";
-          payload: {
-              participantId: string;
-              isAudioEnabled: boolean;
-          };
-      }
-    | {
-          type: "PARTICIPANT_JOINED";
-          payload: {
-              paritipant: RemoteParticipantState;
-          };
-      }
-    | {
-          type: "PARTICIPANT_LEFT";
-          payload: {
-              participantId: string;
-          };
-      }
-    | {
-          type: "PARTICIPANT_STREAM_ADDED";
-          payload: {
-              participantId: string;
-              stream: MediaStream;
-          };
-      }
-    | {
-          type: "PARTICIPANT_VIDEO_ENABLED";
-          payload: {
-              participantId: string;
-              isVideoEnabled: boolean;
-          };
-      }
-    | {
-          type: "PARTICIPANT_METADATA_CHANGED";
-          payload: {
-              participantId: string;
-              displayName: string;
-          };
-      }
-    | {
-          type: "LOCAL_CLIENT_DISPLAY_NAME_CHANGED";
-          payload: {
-              displayName: string;
-          };
-      };
-
-function updateParticipant(
-    remoteParticipants: RemoteParticipantState[],
-    participantId: string,
-    updates: Partial<RemoteParticipantState>
-): RemoteParticipantState[] {
-    const existingParticipant = remoteParticipants.find((p) => p.id === participantId);
-    if (!existingParticipant) {
-        return remoteParticipants;
-    }
-    const index = remoteParticipants.indexOf(existingParticipant);
-
-    return [
-        ...remoteParticipants.slice(0, index),
-        { ...existingParticipant, ...updates },
-        ...remoteParticipants.slice(index + 1),
-    ];
-}
-
-export default function reducer(state: RoomState, action: Action): RoomState {
-    switch (action.type) {
-        case "ROOM_JOINED":
-            return {
-                ...state,
-                localParticipant: action.payload.localParticipant,
-                remoteParticipants: action.payload.remoteParticipants,
-                roomConnectionStatus: "connected",
-            };
-        case "PARTICIPANT_AUDIO_ENABLED":
-            return {
-                ...state,
-                remoteParticipants: updateParticipant(state.remoteParticipants, action.payload.participantId, {
-                    isAudioEnabled: action.payload.isAudioEnabled,
-                }),
-            };
-        case "PARTICIPANT_JOINED":
-            return {
-                ...state,
-                remoteParticipants: [...state.remoteParticipants, action.payload.paritipant],
-            };
-        case "PARTICIPANT_LEFT":
-            return {
-                ...state,
-                remoteParticipants: [...state.remoteParticipants.filter((p) => p.id !== action.payload.participantId)],
-            };
-        case "PARTICIPANT_STREAM_ADDED":
-            return {
-                ...state,
-                remoteParticipants: updateParticipant(state.remoteParticipants, action.payload.participantId, {
-                    stream: action.payload.stream,
-                }),
-            };
-        case "PARTICIPANT_VIDEO_ENABLED":
-            return {
-                ...state,
-                remoteParticipants: updateParticipant(state.remoteParticipants, action.payload.participantId, {
-                    isVideoEnabled: action.payload.isVideoEnabled,
-                }),
-            };
-        case "PARTICIPANT_METADATA_CHANGED":
-            return {
-                ...state,
-                remoteParticipants: [
-                    ...state.remoteParticipants.map((p) =>
-                        p.id === action.payload.participantId ? { ...p, displayName: action.payload.displayName } : p
-                    ),
-                ],
-            };
-        case "LOCAL_CLIENT_DISPLAY_NAME_CHANGED":
-            if (!state.localParticipant) return state;
-            return {
-                ...state,
-                localParticipant: { ...state.localParticipant, displayName: action.payload.displayName },
-            };
-        default:
-            throw state;
-    }
-}
